Tweak standardness test to avoid unrelated reverts
The test was set up such that it would revert due to an ownership issue
if the expected revert didn't happen. It doesn't harm the test, but it's
confusing setup for the goal.

Also, the specified revert message included a period and matched due to
how truffle formats revert exceptions, but the period is not in the
original revert message.